+++ flutter_google_places_sdk_platform_interface/lib/src/types/place_type_filter.dart
@@ -28,8 +28,7 @@ enum PlaceTypeFilter {
   /// COUNTRY
   /// ADMINISTRATIVE_AREA_LEVEL_1
   /// ADMINISTRATIVE_AREA_LEVEL_2
-  REGIONS
-  ;
+  REGIONS;
 
   factory PlaceTypeFilter.fromJson(String name) {
     name = name.toLowerCase();
@@ -58,5 +57,5 @@ enum PlaceTypeFilter {
 }
 
 extension PlaceTypeFilterDescriptor on PlaceTypeFilter {
-  String get value => describeEnum(this);
+  String get value => name;
 }